Support importing shelved files via --changesfile argument
There are definite drawbacks to this.  I currently have no good way to update a local branch with the contents of the shelved changes.  They are instead imported into the p4/master branch (or if detectBranches is in use, the appropriate branch is updated).  But to the rest of the git-p4 system, these git commits will look like they're already submitted to the perforce depot.  Since these shelved changes are temporary, it is best to make a local git branch after the import, then manually move the p4/<branch> ref back to what it was before the git p4 sync command was run.